fix(bmfont): refactor openglfont
