--- game/assets/shaders/bmtext.frag
@@ -1,24 +1,0 @@
-#version 100
-
-#ifdef GL_ES
-precision mediump float;
-#endif
-
-varying vec2 vTexCoord;
-
-uniform sampler2D text;
-uniform vec3 textColor;
-uniform float screenPxRange;
-
-float median(float r, float g, float b) {
-    return max(min(r, g), min(max(r, g), b));
-}
-
-void main() {
-    vec3 msd = texture2D(text, vTexCoord).rgb;
-    float sd = median(msd.r, msd.g, msd.b);
-    float screenPxDistance = screenPxRange * (sd - 0.5);
-    float opacity = clamp(screenPxDistance + 0.5, 0.0, 1.0);
-    gl_FragColor = vec4(textColor, 1.0) * vec4(1.0, 1.0, 1.0, opacity);
-    //    gl_FragColor = vec4(textColor, 1.0) * vec4(1.0, 1.0, 1.0, texture2D(text, vTexCoord).a);
-}
