+++ sponge/src/platform/opengl/openglresourcemanager.cpp
@@ -12,7 +12,6 @@
 #define STB_IMAGE_IMPLEMENTATION
 #include "stb_image.h"
 
-std::unordered_map<std::string, std::shared_ptr<OpenGLBMFont>> OpenGLResourceManager::bmfonts;
 std::unordered_map<std::string, std::shared_ptr<OpenGLFont>> OpenGLResourceManager::fonts;
 std::unordered_map<std::string, std::shared_ptr<OpenGLModel>> OpenGLResourceManager::meshes;
 std::unordered_map<std::string, std::shared_ptr<OpenGLShader>> OpenGLResourceManager::shaders;
@@ -34,26 +33,6 @@ std::shared_ptr<OpenGLFont> OpenGLResourceManager::loadFont(const std::string &p
 
     std::shared_ptr<OpenGLFont> font = loadFontFromFile(path, screenWidth, screenHeight);
     fonts[name] = font;
-
-    return font;
-}
-
-std::shared_ptr<OpenGLBMFont> OpenGLResourceManager::getBMFont(const std::string &name) {
-    assert(!name.empty());
-    return bmfonts.at(name);
-}
-
-std::shared_ptr<OpenGLBMFont> OpenGLResourceManager::loadBMFont(const std::string &path, const std::string &name,
-                                                                int screenWidth, int screenHeight) {
-    assert(!path.empty());
-    assert(!name.empty());
-
-    if (fonts.find(name) != fonts.end()) {
-        return bmfonts[name];
-    }
-
-    std::shared_ptr<OpenGLBMFont> font = loadBMFontFromFile(path, screenWidth, screenHeight);
-    bmfonts[name] = font;
 
     return font;
 }
@@ -137,21 +116,9 @@ std::shared_ptr<OpenGLFont> OpenGLResourceManager::loadFontFromFile(const std::s
                                                                     int screenHeight) {
     assert(!path.empty());
 
-    SPONGE_CORE_INFO("Loading font file: {0}", path);
-
-    auto font = std::make_shared<OpenGLFont>(screenWidth, screenHeight);
-    font->load(path, 24);
-
-    return font;
-}
-
-std::shared_ptr<OpenGLBMFont> OpenGLResourceManager::loadBMFontFromFile(const std::string &path, int screenWidth,
-                                                                        int screenHeight) {
-    assert(!path.empty());
-
     SPONGE_CORE_INFO("Loading BM font file: {0}", path);
 
-    auto font = std::make_shared<OpenGLBMFont>(screenWidth, screenHeight);
+    auto font = std::make_shared<OpenGLFont>(screenWidth, screenHeight);
     font->load(path);
 
     return font;
