+++ sponge/src/platform/opengl/openglresourcemanager.h
@@ -5,7 +5,6 @@
 #include <string>
 #include <unordered_map>
 
-#include "openglbmfont.h"
 #include "openglfont.h"
 #include "openglmodel.h"
 #include "openglshader.h"
@@ -15,9 +14,6 @@ class OpenGLResourceManager {
    public:
     static std::shared_ptr<OpenGLFont> getFont(const std::string &name);
     static std::shared_ptr<OpenGLFont> loadFont(const std::string &path, const std::string &name, int screenWidth,
-                                                int screenHeight);
-    static std::shared_ptr<OpenGLBMFont> getBMFont(const std::string &name);
-    static std::shared_ptr<OpenGLBMFont> loadBMFont(const std::string &path, const std::string &name, int screenWidth,
                                                     int screenHeight);
 
     static std::shared_ptr<OpenGLModel> getMesh(const std::string &name);
@@ -35,13 +31,11 @@ class OpenGLResourceManager {
     OpenGLResourceManager() = default;
 
     static std::shared_ptr<OpenGLFont> loadFontFromFile(const std::string &path, int screenWidth, int screenHeight);
-    static std::shared_ptr<OpenGLBMFont> loadBMFontFromFile(const std::string &path, int screenWidth, int screenHeight);
     static std::shared_ptr<OpenGLModel> loadMeshFromFile(const std::string &path);
     static std::string loadSourceFromFile(const std::string &path);
     static std::shared_ptr<OpenGLTexture> loadTextureFromFile(const std::string &path);
 
     static std::unordered_map<std::string, std::shared_ptr<OpenGLFont>> fonts;
-    static std::unordered_map<std::string, std::shared_ptr<OpenGLBMFont>> bmfonts;
     static std::unordered_map<std::string, std::shared_ptr<OpenGLModel>> meshes;
     static std::unordered_map<std::string, std::shared_ptr<OpenGLShader>> shaders;
     static std::unordered_map<std::string, std::shared_ptr<OpenGLTexture>> textures;
